Work around pycdio crash at end of rip cd info. Check drive missing.
On Darwin (Mac OS X), sometimes libcdio will lose access to CD drive,
if it has just been busy with another task. os.stat(path) tests for
this. If stat returns an OSError #2, the CD drive is inaccessible for
now. On my system, 0.5 seconds wait is long enough for the drive to
reappear. But we have a few longer delays in reserve.